+++ .githooks/update-structure.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-// @subframe-version 0.12.0-beta
+// @subframe-version 0.14.0-beta
 // @subframe-managed
 /**
  * SubFrame STRUCTURE.json Updater